SwiftCompilerSources: split `SIL/Utils.swift` into multiple files
