+++ SwiftCompilerSources/Sources/SIL/Utilities/CMakeLists.txt
@@ -8,6 +8,7 @@
 
 swift_compiler_sources(SIL
   AccessUtils.swift
+  SequenceUtilities.swift
   SmallProjectionPath.swift
   WalkUtils.swift
 )
